WebIDL/tests/submissions/heycam/octet-002.html

<!DOCTYPE html>
<title>octet</title>
<script src="http://w3c-test.org/resources/testharness.js"></script>
<script src="http://w3c-test.org/resources/testharnessreport.js"></script>
<script src="utils.js"></script>
<div id="log"></div>

<script>
// 4.2.5 octet
//
// IDL octet values are represented as integer JS Number values in the range
// [0, 255].

// feature: Uint8Array (indexed property getter) (of type octet)

test(function() {
  var a = new Uint8Array(1);
  assert_equals(typeof a[0], "number");
  assert_true(a[0] == Math.floor(a[0]) && a[0] >= 0 && a[0] <= 255);
}, "Int8Array indexed getter returns a value of the correct type");
</script>
